warnings: fix another batch after enabling /W4 (size_t conversions, safer signatures)
Follow-up to the “enable stricter MSVC warnings” change: address the next wave of diagnostics
reported by /W4 by adding explicit size_t→int casts where we intentionally use signed indices
and counters, plus a few small safety guards for empty inputs to avoid UB.

Also tighten a couple of APIs (pass strings by const reference) and adjust a few helpers to keep
index arithmetic well-defined.

More warning fixes will follow in subsequent commits.

+++ LeetCodeCpp/LongestIncreasingSubsequence_300.cpp
@@ -116,7 +116,7 @@ int LongestIncreasingSubsequence_300::lengthOfLIS_tails(vector<int> &nums)
             *it = num;
         }
     }
-    return tails.size();
+    return static_cast<int>(tails.size());
 }
 
 /*
